# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the pantahub API documentation to use our API.",
        "title": "Pantahub API reference",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/tokens": {
            "get": {
                "description": "List tokens for a owner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tokens"
                ],
                "summary": "List tokens for a owner",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Owner",
                        "name": "owner",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "CreatedAt",
                        "name": "createdAt",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/tokenservice.ListOfToken"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.RError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/utils.RError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.RError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.RError"
                        }
                    }
                }
            }
        },
        "/tokens/": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new token for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tokens"
                ],
                "summary": "Create a new token for a user",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Create Token Request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tokenservice.AuthTokenReqPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tokenmodels.AuthToken"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.RError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/utils.RError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.RError"
                        }
                    }
                }
            }
        },
...
        "/trails": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "devices get a list of one and only one trail. users get trails for all the\ndevices they have trail control over (right now simplified for owner)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trails"
                ],
                "summary": "Get a list of trails",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/trailmodels.Trail"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.RError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.RError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.RError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new trails. usually done by device on first log in.\ninitiates the trail by using the reported state as stepwanted 0 and setting\nthe step 0 to be the POSTED JSON. Either device accounts or user accounts can\ndo this for devices owned, but there can always only be ONE trail per device.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trails"
                ],
                "summary": "Create a new trails",
                "parameters": [
                    {
                        "description": "initial state",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/trails.state"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/trailmodels.Trail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.RError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.RError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.RError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "tokenmodels.AuthToken": {
            "type": "object",
            "properties": {
                "deleted": {
                    "type": "boolean"
                },
                "deleted-at": {
                    "type": "string"
                },
                "expire-at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "owner-nick": {
                    "type": "string"
                },
                "parse-scopes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/utils.Scope"
                    }
                },
                "prn": {
                    "type": "string"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "secret": {
                    "type": "string"
                },
                "time-created": {
                    "type": "string"
                },
                "time-modified": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "tokenservice.AuthTokenReqPayload": {
            "type": "object",
            "properties": {
                "expire-at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "tokenservice.ListOfToken": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tokenmodels.AuthToken"
                    }
                },
                "next": {
                    "type": "string"
                },
                "page_offset": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "prev": {
                    "type": "string"
                },
                "resource": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "trailmodels.Trail": {
            "type": "object",
            "properties": {
                "device": {
                    "type": "string"
                },
                "factory-state": {
                    "type": "object",
                    "additionalProperties": true
                },
                "id": {
                    "type": "string"
                },
                "last-insync": {
                    "description": "Admins   []string `json:\"admins\"`   // XXX: maybe this is best way to do delegating device access....",
                    "type": "string"
                },
                "last-touched": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "used_objects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "trails.state": {
            "type": "object",
            "additionalProperties": true
        },
...
        "utils.RError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "incident": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "utils.Scope": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                },
                "service": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "/auth/authorize",
            "tokenUrl": "/auth/token"
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "/auth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "/auth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/auth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "read": " Grants read access",
                "write": " Grants write access"
...
}