# trails/get-trails.go
//
// Copyright (c) 2017-2026 Pantacor Ltd.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//   http://www.apache.org/licenses/LICENSE-2.0
//
//   Unless required by applicable law or agreed to in writing, software
//   distributed under the License is distributed on an "AS IS" BASIS,
//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
//   See the License for the specific language governing permissions and
//   limitations under the License.
//

// Package trails offer a two party master/slave relationship enabling
// the master to asynchronously deploy configuration changes to its
// slave in a stepwise manner.
package trails

import (
	"fmt"
	"log"
	"net/http"
	"time"

	"context"

	jwtgo "github.com/golang-jwt/jwt/v5"
	"github.com/labstack/echo/v5"
	"gitlab.com/pantacor/pantahub-base/trails/trailmodels"
	"gitlab.com/pantacor/pantahub-base/utils"
	"gitlab.com/pantacor/pantahub-base/utils/echoutil"
	"go.mongodb.org/mongo-driver/mongo/options"
	"gopkg.in/mgo.v2/bson"
)

// handleGetTrails Get a list of trails
// @Summary Get a list of trails
// @Description devices get a list of one and only one trail. users get trails for all the
// @Description devices they have trail control over (right now simplified for owner)
// @Accept  json
// @Produce  json
// @Tags trails
// @Security ApiKeyAuth
// @Success 200 {array} trailmodels.Trail
// @Failure 400 {object} utils.RError
// @Failure 404 {object} utils.RError
// @Failure 500 {object} utils.RError
// @Router /trails [get]
func (a *App) handleGetTrails(c *echo.Context) error {
	fmt.Printf("handleGetTrails\n")

	owner, ok := c.Get(echoutil.KeyJWTPayload).(jwtgo.MapClaims)["prn"]
	if !ok {
		// XXX: find right error
		return echoutil.RestErrorWrapper(c, "You need to be logged in", http.StatusForbidden)
	}

	authType, ok := c.Get(echoutil.KeyJWTPayload).(jwtgo.MapClaims)["type"]

	coll := a.mongoClient.Database(utils.MongoDb).Collection("pantahub_trails")

	if coll == nil {
		return echoutil.RestErrorWrapper(c, "Error with Database connectivity", http.StatusInternalServerError)
	}
	ownerField := ""
	if authType == "DEVICE" {
		ownerField = "device"
	} else if authType == "USER" || authType == "SESSION" {
		ownerField = "owner"
	}

	trails := make([]trailmodels.Trail, 0)

	findOptions := options.Find()
	findOptions.SetNoCursorTimeout(true)
	ctx, cancel := context.WithTimeout(c.Request().Context(), 10*time.Second)
	defer cancel()
	cur, err := coll.Find(ctx, bson.M{
		ownerField: owner,
		"garbage":  bson.M{"$ne": true},
	}, findOptions)
	if err != nil {
		return echoutil.RestErrorWrapper(c, "Error on fetching devices:"+err.Error(), http.StatusForbidden)
	}
	defer cur.Close(ctx)

	for cur.Next(ctx) {
		result := trailmodels.Trail{}
		err := cur.Decode(&result)
		if err != nil {
			return echoutil.RestErrorWrapper(c, "Cursor Decode Error:"+err.Error(), http.StatusForbidden)
		}
		result.FactoryState = utils.BsonUnquoteMap(&result.FactoryState)
		trails = append(trails, result)
	}

	if authType == "DEVICE" {
		if len(trails) > 1 {
			log.Println("WARNING: more than one trail in db for device - bad DB: " + owner.(string))
			trails = trails[0:1]
		}
	}
	return echoutil.WriteJSON(c, http.StatusOK, trails)
}